requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
pandas>=2.2.0
torch>=2.2.0 